requirements.txt
requests==2.22.0
python-dotenv==0.12.0
vk-api==11.8.0
PySocks==1.7.1
python-telegram-bot==12.5.1